--- server/v2/core/README.md
@@ -1,3 +1,0 @@
-# Server Core
-
-The core package in the server directory defines interfaces that are needed to build a custom server.
