--- server/v2/core/go.mod
@@ -1,22 +1,0 @@
-module cosmossdk.io/server/v2/core
-
-go 1.21
-
-replace cosmossdk.io/core => ../../../core
-
-require (
-	cosmossdk.io/core v0.11.0
-	github.com/cosmos/gogoproto v1.4.11
-)
-
-require (
-	cosmossdk.io/log v1.3.1 // indirect
-	github.com/google/go-cmp v0.6.0 // indirect
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.20 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/rs/zerolog v1.32.0 // indirect
-	golang.org/x/exp v0.0.0-20231006140011-7918f672742d // indirect
-	golang.org/x/sys v0.17.0 // indirect
-	google.golang.org/protobuf v1.33.0 // indirect
-)
